server/requirements.txt
# AI &  LLM
langchain
google-generativeai
sentence-transformers
pinecone-client
torch

# Web
fastapi
uvicorn[standard]

# DB/ORM
sqlalchemy 
peewee 
psycopg2-binary

# Util/Auth
python-dotenv
passlib[bcrypt]
python-jose[cryptography]